requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.6
aiosignal==1.3.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
blinker==1.8.2
click==8.1.7
cypari==2.5.5
decorator==5.1.1
dnspython==2.6.1
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.16.1
Flask==3.0.3
frozenlist==1.4.1
FXrays==1.3.5
idna==3.10
ipython==8.27.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==0.10.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kafka-python==2.0.2
knot-floer-homology==1.2
low-index==1.2
lz4==4.3.3
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mlflow==2.17.2
msgpack==1.1.0
multidict==6.1.0
mysql-connector-python==9.0.0
networkx==3.3
numpy==2.1.1
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
plink==2.4.2
prompt_toolkit==3.0.48
propcache==0.2.0
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.7
Pygments==2.18.0
pymongo==4.9.1
pypng==0.20220715.0
pyspark==3.5.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
ray==2.38.0
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.8.1
rich-argparse==1.5.2
rpds-py==0.20.0
scikit-learn==1.5.2
scipy==1.14.1
setproctitle==1.3.3
six==1.16.0
snappy==3.1.1
snappy-manifolds==1.2
spherogram==2.2.1
stack-data==0.6.3
tabulate==0.9.0
tenacity==8.5.0
termcolor==2.4.0
text-unidecode==1.3
threadpoolctl==3.5.0
time-machine==2.15.0
tomli==2.0.2
tqdm==4.66.6
traitlets==5.14.3
tzdata==2024.2
wcwidth==0.2.13
Werkzeug==3.0.4
yarl==1.15.3
zstandard==0.23.0
coverage
joblib
kafka-python-ng
